requirements.txt
fastapi
uvicorn
openai
pydantic
requests
pandas
numpy
scipy
python-dotenv
